build: refactored setup scripts to src/ .ts files
