+++ tsconfig.json
@@ -6,10 +6,11 @@
 		"module": "esnext",
 		"moduleResolution": "node",
 		"outDir": "lib",
+		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"sourceMap": true,
 		"strict": true,
-		"target": "ES2021"
+		"target": "ES2022"
 	},
 	"include": ["src"]
 }
